python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==1.95.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.6.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
